requirements.txt
aiogram==3.4.1
SQLAlchemy>=1.4
asyncpg
aiosqlite
python-dotenv
